Add machine which generates a random melody
This required refactoring the way machines are created, to pass in a context object,  with a random number generator. This can be extended in the future to give machines access to other useful information. Oh and BTW I rewrote Params to allow different Param types to be added more easily.