Almost a rewrite...
Introduce a new kind of machine which generates a random sequence...
Which meant rewriting the Sequencer/Track/Machine structure to allow an
RNG to be passed...
Which required some conditional compilation to provide access to
platform-specific RNGs...
Which didn't require rewriting params, but honestly I had started that
first and couldn't face unpicking the changes.

Probably other stuff I've forgotten.